+++ NEWS.md
@@ -6,6 +6,14 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v0.14.2
+
+- Fixes
+
+  - Fix Librarian process hanging due to logging deadlock.
+  - Fix reader keyboard shortcut help.
+  - Fix book change drawer appearing in the middle of books.
+
 ## v0.14.1
 
 - Fixes